Significant speedup of amber module
If not running in debug mode, the md module reduces
all access to the filesystem to a minimum by sending
MD output to /dev/null and creating tempfiles on the
local disk for cpptraj_in and cpptraj_output files.
This will significantly speed up hdWE especially in
a multinode NFS environment.