Combinatoric bins
Bins are created, based on the are in the ramachandran
plot of (phi,psi) angles per residue. The mapping function
between (phi,psi) and the bin index returns indices 1,2 or 3
depending on the area alpha, beta and rest. The program
hardcodes the simulation of a 10 residue alanine helix so far.
Only for testing!