Error handling in amber_module
Some error handling of file stale errors and
PMEMD STOP is done by calling os.system('sync')
and sleeping for 2 seconds. Most of the errors
are filesystem related and can be solved by syncing
and waiting a bit before retrying to read input files.
This way is still far better that just crashing...