Split forward merge mode
In order to allow for better sampling in a specific
coordinate direction, the 'split-forward' merge mode
splits all segments which move in the desired direction
(specified by 'split-forward-coordinate') N times where
N is given by the config entry 'split-forward-children'.
The split forward merge mode is therefore not a classical
merge mode as no segments are merged.