Make start-bin-coordinate-ids an optional
The input file parameter start-bin-coordinate-ids
can be used to only select starting structures which reside
in the given bin ids when setting up a simulation. However,
when the user wants all his input structures to be used for the
WE setup, he should not be asked to specify this parameter.